# src/main/java/dk/sdu/mmmi/jobservice/service/model/Job.java
package dk.sdu.mmmi.jobservice.service.model;

import jakarta.persistence.*;
import lombok.Getter;
import lombok.Setter;
import lombok.ToString;

import java.util.Date;

@Entity
@Table(name = "jobs")
@Getter
@Setter
@ToString
public class Job {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @Column(name = "title", nullable = false)
    private String title;

    @Column(name = "location", nullable = false)
    private String location;

    @Column(name = "description", nullable = false)
    private String description;

    @Column(name = "job_type", nullable = false)
    private String jobType;

    @Column(name = "salary")
    private Double salary;

    @Column(name = "created_at", nullable = false)
    private Date createdAt;

    @Column(name = "updated_at", nullable = false)
    private Date updatedAt;

    @Column(name = "expires_at", nullable = false)
    private Date expiresAt;

    @Column(name = "company", nullable = false)
    private String company;
}